// src/graphql/__generated__/UserInfo.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: UserInfo
// ====================================================

export interface UserInfo_user {
  __typename: "User";
  displayName: string;
  avatar: string;
}

export interface UserInfo {
  user: UserInfo_user;
}